+++ test/unit/public_suffix_test.rb
@@ -44,7 +44,7 @@ class PublicSuffixTest < Minitest::Unit::TestCase
     assert_equal "one.two", domain.trd
   end
 
-  def test_self_parse_a_fully_qualified_domain_name
+  def test_self_parse_name_fqdn
     domain = PublicSuffix.parse("www.example.com.")
     assert_instance_of PublicSuffix::Domain, domain
     assert_equal "com",     domain.tld
@@ -67,27 +67,31 @@ class PublicSuffixTest < Minitest::Unit::TestCase
     end
   end
 
-  def test_self_parse_a_domain_with_custom_list
+  def test_self_parse_with_custom_list
     list = PublicSuffix::List.new
     list << PublicSuffix::Rule.factory("test")
 
     domain = PublicSuffix.parse("www.example.test", list)
+    assert_instance_of PublicSuffix::Domain, domain
     assert_equal "test",    domain.tld
     assert_equal "example", domain.sld
     assert_equal "www",     domain.trd
   end
 
-  def test_self_parse_raises_with_invalid_domain
-    error = assert_raises(PublicSuffix::DomainInvalid) { PublicSuffix.parse("example.qqq") }
-    assert_match %r{example\.qqq}, error.message
+  def test_self_parse_with_notlisted_name
+    domain = PublicSuffix.parse("example.tldnotlisted")
+    assert_instance_of PublicSuffix::Domain, domain
+    assert_equal "tldnotlisted",    domain.tld
+    assert_equal "example",         domain.sld
+    assert_equal nil,               domain.trd
   end
 
-  def test_self_parse_raises_with_unallowed_domain
+  def test_self_parse_with_unallowed_domain
     error = assert_raises(PublicSuffix::DomainNotAllowed) { PublicSuffix.parse("example.ke") }
     assert_match %r{example\.ke}, error.message
   end
 
-  def test_self_raises_with_uri
+  def test_self_parse_with_uri
     error = assert_raises(PublicSuffix::DomainInvalid) { PublicSuffix.parse("http://google.com") }
     assert_match %r{http://google\.com}, error.message
   end
@@ -100,17 +104,16 @@ class PublicSuffixTest < Minitest::Unit::TestCase
     assert  PublicSuffix.valid?("www.google.co.uk")
   end
 
-  # Returns false when domain has an invalid TLD
-  def test_self_valid_with_invalid_tld
-    assert !PublicSuffix.valid?("google.qqq")
-    assert !PublicSuffix.valid?("www.google.qqq")
+  def test_self_valid_with_notlisted_name
+    assert  PublicSuffix.valid?("google.tldnotlisted")
+    assert  PublicSuffix.valid?("www.google.tldnotlisted")
   end
 
-  def test_self_valid_with_fully_qualified_domain_name
-    assert  PublicSuffix.valid?("google.com.")
-    assert  PublicSuffix.valid?("google.co.uk.")
-    assert !PublicSuffix.valid?("google.qqq.")
-  end
+  # def test_self_valid_with_fully_qualified_domain_name
+  #   assert  PublicSuffix.valid?("google.com.")
+  #   assert  PublicSuffix.valid?("google.co.uk.")
+  #   assert !PublicSuffix.valid?("google.tldnotlisted.")
+  # end
 
 
   def test_self_domain
@@ -120,17 +123,54 @@ class PublicSuffixTest < Minitest::Unit::TestCase
     assert_equal "google.co.uk",  PublicSuffix.domain("www.google.co.uk")
   end
 
-  def test_self_domain_with_invalid_domain_returns_nil
-    assert_nil PublicSuffix.domain("example.qqq")
+  def test_self_domain_with_notlisted_name
+    assert_equal "example.tldnotlisted", PublicSuffix.domain("example.tldnotlisted")
   end
 
-  def test_self_domain_with_unallowed_domain_returns_nil
+  def test_self_domain_with_unallowed_name
     assert_nil PublicSuffix.domain("example.ke")
   end
 
-  def test_self_domain_with_blank_sld_returns_nil
+  def test_self_domain_with_blank_sld
     assert_nil PublicSuffix.domain("com")
     assert_nil PublicSuffix.domain(".com")
+  end
+
+
+  def test_self_normalize
+    [
+        ['com', 'com'],
+        ['example.com', 'example.com'],
+        ['www.example.com', 'www.example.com'],
+
+        ['example.com.',  'example.com'],     # strip FQDN
+        [' example.com ', 'example.com'],     # strip spaces
+        ['Example.COM',   'example.com'],     # downcase
+    ].each do |input, output|
+      assert_equal output, PublicSuffix.normalize(input)
+    end
+  end
+
+  def test_normalize_blank
+    [
+        nil,
+        '',
+        ' '
+    ].each do |input, output|
+      error = PublicSuffix.normalize(input)
+      assert_instance_of PublicSuffix::DomainInvalid, error
+      assert_equal "Name is blank", error.message
+    end
+  end
+
+  def test_normalize_scheme
+    [
+        'https://google.com'
+    ].each do |input, output|
+      error = PublicSuffix.normalize(input)
+      assert_instance_of PublicSuffix::DomainInvalid, error
+      assert_match /scheme/, error.message
+    end
   end
 
 end