+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - CHANGED: Removed deprecated PublicSuffix::InvalidDomain exception
 
+- CHANGED: If the suffix is now listed, then the prevaling rule is "*" as defined by the PSL algorithm (GH-91)
+
+- CHANGED: Input validation is performed only if you call PublicSuffix.parse or PublicSuffix.list
+
 
 #### Release 1.5.3